renamed:    version.h -> version_compat.h
