+++ version_compat.h
